Refactor Merkle tree
This commit completely refactors the Merkle tree class and its subclasses.
The code before was difficult to follow and pretty buggy, so this attempts
to turn the approach around, hopefully fixing many issues and improving
my own sanity. However, some items are not yet complete so there's more
to do here.